+++ src/main/java/org/example/repositories/ShowRepository.java
@@ -3,4 +3,4 @@ import org.example.entities.Show;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 @Repository
-public interface ShowRepository extends JpaRepository<Show, Integer> {}
+public interface ShowRepository extends JpaRepository<Show, Long> {}
